Disabled The One Ring
The One Ring is beyond OP. Your only punishment is warp if you wear it a while... That being said it gives you Invisibility and invulnerability that lets you really cheat things in the game. These have a 15% chance in dungeon loot and I think it's silly.

There is a quest reward for it but it's on its own and as far as I know hidden.

So I think it's best we hide this and forget it unless it becomes a must have item later and I don't think it does.