Server Readme: Lots of Quest Stuff
So we are now disabling a few quests with no continuation that might be server / single player info meant for admins as well as things like the sleeping bag quest. It just makes it easier then replacing things now we just add a simple number to them that gets locked behind one the most end game items lol.